feat: Conditionally create API Gateway
Feature flag controls creation of resources as well as use thereof.
